typo

+++ meta-negative-rep.py
@@ -20,7 +20,7 @@ parser = argparse.ArgumentParser(description='Pytorch Implementation of Neural P
 parser.add_argument('--batch_size', default=256, type=int)
 parser.add_argument('--epochs', default=1000, type=int)
 parser.add_argument('--lr', default=1e-3, type=float)
-parser.add_argument('--temp', default=0.5, type=int)
+parser.add_argument('--temp', default=0.5, type=float)
 parser.add_argument('--download', default=False, type=bool)
 
 args = parser.parse_args()
